[✨feat] 투표자 한 명 삭제와 체크박스로 투표자 선택 구현(VOTEGUARD-66)
- 투표자 한 명을 등록할 때 뜨는 창을 따로 컴포넌트 분리
- 체크박스를 통해 등록된 투표자 전체를 선택할 수 있고 하나를 따로 선택하는 것도 가능하도록 구현
- 체크박스로 선택한 투표자의 인원 수 표시
